+++ src/app/dashboard/counter/page.tsx
@@ -1,4 +1,4 @@
-import { CartCounter } from "@/app/shopping-cart";
+import { CartCounter } from "@/shopping-cart";
 
 export const metadata = {
 	title: "Shopping Cart",
