+++ src/app/dashboard/layout.tsx
@@ -1,4 +1,4 @@
-import { SideBar } from "../ components";
+import { SideBar } from "@/ components";
 
 export default function DashboardLayout({
 	children,
