+++ src/app/dashboard/pokemons/page.tsx
@@ -1,4 +1,4 @@
-import { PokemonGrid, PokemonsResponse, SimplePokemon } from "@/app/pokemons";
+import { PokemonGrid, PokemonsResponse, SimplePokemon } from "@/pokemons";
 import { notFound } from "next/navigation";
 
 const getPokemons = async (
