+++ src/pokemons/interfaces/simple-pokemons.ts
