+++ src/shopping-cart/components/CartCounter.tsx
