+++ tailwind.config.ts
@@ -1,11 +1,14 @@
 import type { Config } from "tailwindcss";
 
 const config: Config = {
-  content: [
-    "./src/pages/**/*.{js,ts,jsx,tsx,mdx}",
-    "./src/components/**/*.{js,ts,jsx,tsx,mdx}",
-    "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
-  ],
+  content: ["./src/**/*.{js,ts,jsx,tsx,mdx}"],
+  // content: [
+  //   './src/pages/**/*.{js,ts,jsx,tsx,mdx}',
+  //   './src/components/**/*.{js,ts,jsx,tsx,mdx}',
+  //   './src/app/**/*.{js,ts,jsx,tsx,mdx}',
+  //   './src/pokemons/**/*.{js,ts,jsx,tsx,mdx}',
+  //   './src/shopping-cart/**/*.{js,ts,jsx,tsx,mdx}',
+  // ],
   theme: {
     extend: {
       backgroundImage: {
